backend/requirements.txt
flask
flask-cors
transformers
torch
validators
newspaper3k